pyproject.toml
[project]
name = "expense-tracker"
version = "0.1.0"
description = "Personal Expense Tracker"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.13.3",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.4",
    "greenlet>=3.1.1",
    "passlib==1.7.4",
    "pydantic[email]>=2.9.2",
    "pydantic-settings>=2.6.1",
    "python-dotenv>=1.0.1",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.17",
    "sqlalchemy>=2.0.36",
    "uvicorn>=0.32.0",
    "bcrypt==4.0.1",
    "psycopg2-binary>=2.9.10",
]

[tool.uv]
dev-dependencies = [
    "black>=24.10.0",
    "flake8>=7.1.1",
    "httpx>=0.27.2",
    "isort>=5.13.2",
    "pytest>=8.3.3",
    "pytest-cov>=6.0.0",
    "ruff>=0.7.2",
]
